fix(l1): preserve staffing lifecycle invariants
